+++ internal/dcache_lib/meta_manager/meta_manager.go
@@ -31,7 +31,7 @@
    SOFTWARE
 */
 
-package dcachelib
+package meta_manager
 
 // MetaFile represents the metadata structure for a file
 type MetaFile struct {
@@ -61,47 +61,4 @@ type MetaManager interface {
 
 	// GetFileContent reads and returns the content of a file
 	GetFileContent(filename string) ([]byte, error)
-}
-
-// FileMetaManager is the implementation of MetaManager interface
-type FileMetaManager struct {
-	cacheDir string
-}
-
-// NewMetaManager creates a new implementation of the MetaManager interface
-func NewMetaManager(cacheID string) (MetaManager, error) {
-	return &FileMetaManager{
-		cacheDir: cacheID,
-	}, nil
-}
-
-// Implement all interface methods
-func (m *FileMetaManager) CreateMetaFile(filename string, mvList []string) error {
-	// Implementation here
-	return nil
-}
-
-func (m *FileMetaManager) DeleteMetaFile(filename string) error {
-	// Implementation here
-	return nil
-}
-
-func (m *FileMetaManager) IncrementHandleCount(filename string) error {
-	// Implementation here
-	return nil
-}
-
-func (m *FileMetaManager) DecrementHandleCount(filename string) error {
-	// Implementation here
-	return nil
-}
-
-func (m *FileMetaManager) GetHandleCount(filename string) (int64, error) {
-	// Implementation here
-	return 0, nil
-}
-
-func (m *FileMetaManager) GetFileContent(filename string) ([]byte, error) {
-	// Implementation here
-	return nil, nil
 }
