+++ internal/dcache_lib/meta_manager/meta_manager_impl.go
@@ -1,0 +1,44 @@
+package meta_manager
+
+// FileMetaManager is the implementation of MetaManager interface
+type FileMetaManager struct {
+	cacheDir string
+}
+
+// NewMetaManager creates a new implementation of the MetaManager interface
+func NewMetaManager(cacheID string) (MetaManager, error) {
+	return &FileMetaManager{
+		cacheDir: cacheID,
+	}, nil
+}
+
+// Implement all interface methods
+func (m *FileMetaManager) CreateMetaFile(filename string, mvList []string) error {
+	// Implementation here
+	return nil
+}
+
+func (m *FileMetaManager) DeleteMetaFile(filename string) error {
+	// Implementation here
+	return nil
+}
+
+func (m *FileMetaManager) IncrementHandleCount(filename string) error {
+	// Implementation here
+	return nil
+}
+
+func (m *FileMetaManager) DecrementHandleCount(filename string) error {
+	// Implementation here
+	return nil
+}
+
+func (m *FileMetaManager) GetHandleCount(filename string) (int64, error) {
+	// Implementation here
+	return 0, nil
+}
+
+func (m *FileMetaManager) GetFileContent(filename string) ([]byte, error) {
+	// Implementation here
+	return nil, nil
+}
